Yet more norming
